Cleanup typos and fix flushing
This commit does a few things actually. It cleans up the usage of
`reply.fail` to be exclusively `reply.fail!`. After reviewing my usage,
I realized every failure was killable at that time.

This commit also fixes the flushing of the status lock file. Now the
status can be tracked over time. I need to play with the lock file a
little also. Currently I uses timestamps, but I'm starting to think an
MD5 of the dest or a pid might make more sense and be easier to work
with. I should probably track the pid one way or another so I can kill
operations if needed.