Remove test from absolute_path_validator
Without this commit the validator incorrectly uses the `Pathname` class
to check if the path being validated currently an actual file on the
system the validator is currently running on. Not only does this not
validate the path correctly, it imposes stricter rules than a validator
for an 'absolute path' should. Directories can be absolute too, and are
a path, right?

This commit removes that test.